move "last-screen" handling out of generic code
