+++ subiquity/core.py
@@ -326,8 +326,13 @@ class Subiquity(TuiApplication):
                 print()
                 break
 
+    def load_serialized_state(self):
+        for controller in self.controllers.instances:
+            controller.load_state()
+
     async def start(self):
         self.controllers.load_all()
+        self.load_serialized_state()
         await self.connect()
         if self.opts.autoinstall is not None:
             await self.load_autoinstall_config()
@@ -346,6 +351,15 @@ class Subiquity(TuiApplication):
             self.show_error_report(exc.error_report_ref)
             return
         super()._exception_handler(loop, context)
+
+    def _remove_last_screen(self):
+        last_screen = self.state_path('last-screen')
+        if os.path.exists(last_screen):
+            os.unlink(last_screen)
+
+    def exit(self):
+        self._remove_last_screen()
+        super().exit()
 
     def extra_urwid_loop_args(self):
         return dict(input_filter=self.input_filter.filter)
@@ -411,6 +425,20 @@ class Subiquity(TuiApplication):
         if isinstance(self.ui.body, BaseView):
             self.ui.body.remove_overlay(overlay)
 
+    def initial_controller_index(self):
+        if not self.updated:
+            return 0
+        state_path = self.state_path('last-screen')
+        if not os.path.exists(state_path):
+            return 0
+        with open(state_path) as fp:
+            last_screen = fp.read().strip()
+        controller_index = 0
+        for i, controller in enumerate(self.controllers.instances):
+            if controller.name == last_screen:
+                controller_index = i
+        return controller_index
+
     def select_initial_screen(self, index):
         self.error_reporter.load_reports()
         for report in self.error_reporter.reports:
@@ -454,6 +482,8 @@ class Subiquity(TuiApplication):
             view = await super().make_view_for_controller(new)
             if new.answers:
                 self.aio_loop.call_soon(new.run_answers)
+            with open(self.state_path('last-screen'), 'w') as fp:
+                fp.write(new.name)
             return view
         else:
             if self.autoinstall_config and not new.autoinstall_applied:
